requirements.txt
pyserial==3.4
PySide2==5.15.0
